Merging project dev to main

+++ js/app.js
@@ -1,5 +1,3 @@
-
-
 let tasks = JSON.parse(localStorage.getItem("tasks")) || [];
 
 document.getElementById("taskForm").addEventListener("submit", (e) => {
@@ -12,6 +10,7 @@ document.getElementById("taskForm").addEventListener("submit", (e) => {
   localStorage.setItem("tasks", JSON.stringify(tasks));
 
   document.getElementById("taskForm").reset();
+  renderTasks();
 });
 
 function renderTasks() {
@@ -20,27 +19,38 @@ function renderTasks() {
 
   tasks.forEach(t => {
     const div = document.createElement("div");
-    div.textContent = t.text;
+
+    div.innerHTML = `
+      <span>${t.text}</span>
+
+      <div class="buttons">
+        <button class="edit-btn" onclick="editTask(${t.id})">Editar</button>
+        <button class="delete-btn" onclick="deleteTask(${t.id})">Eliminar</button>
+      </div>
+    `;
+
     container.appendChild(div);
   });
 }
+
 function deleteTask(id) {
   tasks = tasks.filter(t => t.id !== id);
   localStorage.setItem("tasks", JSON.stringify(tasks));
   renderTasks();
 }
 
-
 function editTask(id) {
-  const newText = prompt("Edita la tarea:");
+  const task = tasks.find(t => t.id === id);
 
-  tasks = tasks.map(t => 
+  const newText = prompt("Edita la tarea:", task.text);
+  if (!newText) return;
+
+  tasks = tasks.map(t =>
     t.id === id ? { ...t, text: newText } : t
   );
 
   localStorage.setItem("tasks", JSON.stringify(tasks));
   renderTasks();
 }
-
 
 renderTasks();
